requirements.txt
pandas
great_expectations
gcsfs
google-cloud-storage
python-dotenv
matplotlib
seaborn
plotly
geopandas
nbformat
featuretools
setuptools
scikit-learn
xgboost
fastparquet
optuna
shap
imblearn
dvc